UI: save long plan to scene

+++ selfdrive/ui/paint.cc
@@ -1533,7 +1533,7 @@ static void ui_draw_measures(UIState *s){
           case UIMeasure::VISION_VF: 
             {
               snprintf(name, sizeof(name), "V: VF");
-              snprintf(val, sizeof(val), "%.2f", scene.longitudinal_plan.getVisionVf());
+              snprintf(val, sizeof(val), "%.2f", (float)scene.longitudinal_plan.getVisionVf());
             }
             break;
